# os/arch/arm/src: SMP/AMP migrate systick control to running core during pause