--- a/os/arch/arm/src/amebasmart/amebasmart_timerisr.c
+++ b/os/arch/arm/src/amebasmart/amebasmart_timerisr.c
@@ -60,6 +60,10 @@
 #  error SYSTICK_RELOAD exceeds the range of the RELOAD register
 #endif
 
+#if defined(CONFIG_SMP) && CONFIG_SMP_NCPUS > 1
+volatile uint32_t g_active_system_timer_cpu;
+#endif
+
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
@@ -75,8 +79,14 @@
 
 int up_timerisr(int irq, uint32_t *regs)
 {
-    /* Process timer interrupt */
+#if defined(CONFIG_SMP) && CONFIG_SMP_NCPUS > 1
+    /* Process timer interrupt, only if the active timer is set to this cpu's PPI */
+    if (g_active_system_timer_cpu == up_cpu_index()) {
+      sched_process_timer();
+    }
+#else
     sched_process_timer();
+#endif
 
     /* Clear interrupt */
     uint32_t delta_ticks;