--- a/os/arch/arm/src/armv7-a/arm_cpupause.c
+++ b/os/arch/arm/src/armv7-a/arm_cpupause.c
@@ -52,8 +52,11 @@
 #include "up_internal.h"
 #include "gic.h"
 #include "sched/sched.h"
+#include "arch_timer.h"
 
 #ifdef CONFIG_SMP
+
+extern volatile uint32_t g_active_system_timer_cpu;
 
 /****************************************************************************
  * Private Data
@@ -363,6 +366,9 @@ int up_cpu_pause(int cpu)
 	spin_lock(&g_cpu_paused[cpu]);
 	spin_unlock(&g_cpu_paused[cpu]);
 
+	/* after taking the spinlocks, this cpu is the active timer, since the other one will be paused */
+	g_active_system_timer_cpu = this_cpu();
+
 	/* On successful return g_cpu_wait will be locked, the other CPU will be
 	 * spinning on g_cpu_wait and will not continue until g_cpu_resume() is
 	 * called.  g_cpu_paused will be unlocked in any case.
@@ -418,6 +424,11 @@ int up_cpu_resume(int cpu)
 		spin_lock(&g_cpu_resumed[cpu]);
 
 		spin_unlock(&g_cpu_resumed[cpu]);
+
+		/* transfer systick control back to cpu0 if cpu0 is being resumed */
+		if (cpu == 0) {
+			g_active_system_timer_cpu = cpu;
+		}
 	}
 
 	return OK;