--- a/os/arch/arm/src/armv7-a/arm_cpustart.c
+++ b/os/arch/arm/src/armv7-a/arm_cpustart.c
@@ -54,6 +54,7 @@
 #include "gic.h"
 #include "sched/sched.h"
 #include "smp.h"
+#include "arch_timer.h"
 
 #ifdef CONFIG_SMP
 
@@ -110,6 +111,9 @@ int arm_start_handler(int irq, void *context, void *arg)
 	struct tcb_s *tcb = this_task();
 
 	svdbg("CPU%d Started\n", this_cpu());
+
+	/* start generic timer PPI for this core */
+	up_timer_initialize();
 
 #ifdef CONFIG_SCHED_INSTRUMENTATION
 	/* Notify that this CPU has started */