This PR demonstrates functional switching of systick count when pausing a core.

As systick is counted by each individual core's timer PPI, it will trigger independently of each other. Currently, this will cause the system tick to have multiple count and affect OS timing.

To avoid multiple counting, only one core will be allowed to increment the common OS tick counter

If Core0 happens to be paused, then the other core should take over, and then return control when Core0 resumes

Please check with flat_dev_ddr config