@edwakuwaku 2025-05-08
why do we need to pass the tickisr to secondary core when primary core is paused? Isn't the tick will be compensated during next cycle of tickisr happens (ie. after primary core is resumed). Also, current implementation will cause resource wastage, since h/w has to always react to secondary cpu tickisr but nothing is being done in f/w.
In https://github.com/Samsung/TizenRT/blob/master/os/arch/arm/src/armv7-a/arm_gicv2.c#L151, all the cores bringup will do this, which all the irq for SGI and PPI will be assigned, I think it should be possible to explicitly disable specific irq num for specific core through the GIC cpu interface. Just need to figure out the correct offset for bit masking through https://github.com/Samsung/TizenRT/blob/master/os/arch/arm/src/armv7-a/gic.h...