@edwakuwaku 2025-05-08
It's clear to me for point 1.
For point 2, the reason to ask this is because, the current implementation has direct interaction between bsp and arch layer code which shouldn't be allowed in such hierarchal OS. Some interface should be created from arch layer, and then invoke from bsp layer. 
Maybe @sunghan-chang can comment this better...