@sunghan-chang 2025-05-09
@kishore-sn @lhenry-realtek @edwakuwaku If we set the irq affinity to both (any core) instead of this PR, what happen?