@edwakuwaku 2025-05-09
> @kishore-sn @lhenry-realtek @edwakuwaku If we set the irq affinity to both (any core) instead of this PR, what happen?

In gic, there are distributor and cpu based interface, for shared peripheral interface, it is controlled through distributor, for SGI and PPI, it is through cpu based interface, each core has its own independent set of registers... I don't know what do you mean by what will happen, it is based on user control to decide the behaviour...