@sunghan-chang 2025-05-11
@kishore-sn @lhenry-realtek It looks we should merge this ASAP. Could you check it ASAP?

> Agreed that this is just a "quick hack" to demo the flow, there should be a proper api to handle the systicks and timers that's not just via direct access of extern variable
> 
> Maybe @kishore-sn has some suggestion to modify the system timer APIs for this, and also to handle more than 2 core situation