@kishore-sn 2025-05-14
> it is using up_cpu

I meant usleep in all cases.