@kishore-sn 2025-05-14
> Hi all, We need following cases to work with this PR:
> 
> 1. cpu1 pauses cpu0. cpu1 goes into sleep. Cpu1 must wake up after sleep duration.
> 2. Cpu1 gates cpu0. Cpu1 goes to sleep. After sleep duration Cpu1 must wake up and continue.
> 3. During normal operation, the system timer must count properly. If we call sleep(30), then wake up must happen after 30 secs.
> 4. cpu1 calls cpu pause for cpu0, at that time, system gets assert.
> 5. cpu1 calls cpu gating for cpu0, at that time system get assert.
> 6. Large flash access operation is happening on cpu1 (so cpu0 gets gated). After the flash operation, the system time must be proper.
> 
> When I checked with this PR,
> 
> * cases 1, 3, 4 and 5 are passing
> * case 2 is still failing. When cpu0 goes into gating and cpu1 goes to sleep, then system gets hanged.
> * I am yet to verify case 6. I will update in sometime.

Case 6 is also failing. 
Test procedure: 
1. Run below steps on cpu1
2. Allocate 5mb buffer with dummy data
3. Open file /mnt/test.txt
4. Write 5mb buffer to file
5. Measure time taken using clock_gettime API

Actual time taken is around 50 to 60secs. But time calculated using clock_gettime is 10secs.