@kishore-sn 2025-05-14
> Agreed that this is just a "quick hack" to demo the flow, there should be a proper api to handle the systicks and timers that's not just via direct access of extern variable
> 
> Maybe @kishore-sn has some suggestion to modify the system timer APIs for this, and also to handle more than 2 core situation

For more than 2 cores scenario, I am not sure how pause- resume will work. Is it possible for core A to pause core B and then core C to resume core B? 
In any case, I think it is better to do currently implementation assuming two cores only. 
However, in order to keep the code of arch layer separate from bsp, I suggest below method. This method does not use global variable. There will be no changes in amebasmart_timerisr.c.

```
diff --git a/os/arch/arm/src/armv7-a/arm_cpupause.c b/os/arch/arm/src/armv7-a/arm_cpupause.c
index fcee79f1f..970ea2086 100644
--- a/os/arch/arm/src/armv7-a/arm_cpupause.c
+++ b/os/arch/arm/src/armv7-a/arm_cpupause.c
@@ -52,6 +52,7 @@
 #include "up_internal.h"
 #include "gic.h"
 #include "sched/sched.h"
+#include "arch_timer.h"

 #ifdef CONFIG_SMP

@@ -363,6 +364,14 @@ int up_cpu_pause(int cpu)
        spin_lock(&g_cpu_paused[cpu]);
        spin_unlock(&g_cpu_paused[cpu]);

+       /* Check if we are pausing cpu0.
+        * In this case, after pause request is being handled by cpu0,
+        * we need to enable timer irq on current cpu
+        */
+       if (cpu == 0) {
+               up_timer_enable();
+       }
+
        /* On successful return g_cpu_wait will be locked, the other CPU will be
         * spinning on g_cpu_wait and will not continue until g_cpu_resume() is
         * called.  g_cpu_paused will be unlocked in any case.
@@ -418,6 +427,11 @@ int up_cpu_resume(int cpu)
                spin_lock(&g_cpu_resumed[cpu]);

                spin_unlock(&g_cpu_resumed[cpu]);
+
+               /* transfer systick control back to cpu0 if cpu0 is being resumed */
+               if (cpu == 0) {
+                       up_timer_disable();
+               }
        }

        return OK;
```