@kishore-sn 2025-05-19
@lhenry-realtek 
There seems to be some issue in your implementation in the gating file. I modified it and now the timing values in both cpus are coming properly. Below are the test results. (5MB file size and 6144KB partition size)
This test is performed with CONFIG_PM=y, but pm_resume not called. So, PM is not active. 

```
TASH>>hello 1
TASH>>Hello, World!!
Running file write test on **cpu 1**
BEFORE: (1747612804 sec, 102000000 nsec)
AFTER:  (1747612815 sec, 713000000 nsec)
Time taken = 11 secs           <<<<<<<<<<<<< Fresh file

TASH>>hello 1
TASH>>Hello, World!!
Running file write test on **cpu 1**
BEFORE: (1747612855 sec, 424000000 nsec)
AFTER:  (1747613512 sec, 45000000 nsec)
Time taken = 657 secs        <<<<<<<<<<<<<<< Overwrite file


TASH>>hello 0
TASH>>Hello, World!!
Running file write test on **cpu 0**
BEFORE: (1747612871 sec, 180000000 nsec)
AFTER:  (1747612882 sec, 818000000 nsec)
Time taken = 11 secs        <<<<<<<<<<<<< Fresh file

TASH>>hello 0
TASH>>Hello, World!!
Running file write test on **cpu 0**
BEFORE: (1747612887 sec, 820000000 nsec)
AFTER:  (1747613548 sec, 76000000 nsec)
Time taken = 661 secs         <<<<<<<<<<<<<<< Overwrite file

```

Please find my changes in the below patch. 
```
diff --git a/os/arch/arm/src/armv7-a/arm_cpugating.c b/os/arch/arm/src/armv7-a/arm_cpugating.c
index 4b3732fdf..753bb7049 100644
--- a/os/arch/arm/src/armv7-a/arm_cpugating.c
+++ b/os/arch/arm/src/armv7-a/arm_cpugating.c
@@ -53,6 +53,10 @@ void up_set_gating_flag_status(uint32_t CoreID, uint32_t val)
        ARM_DSB();
        /* Flag already reach 0 */
        if (!g_cpugating_flag[CoreID]) {
+                if (this_cpu() != 0) {
+                        up_timer_disable();
+                }
+
                SP_SEV();
        }
 }
@@ -75,11 +79,6 @@ void up_do_gating(void)
                        SP_WFE();
                }
                irqrestore(PrevIrqStatus);
-
-               /* transfer systick control back to cpu0 if cpu0 exit gating. cpu1 do not need */
-               if (cpu == 0) {
-                       up_timer_enable();
-               }
        }
 
 }
@@ -124,10 +123,10 @@ void up_cpu_gating(int cpu)
        /* Fire SGI for cpu to enter gating */
        arm_cpu_sgi(GIC_IRQ_SGI3, (1 << cpu));
 
-       /* after gating other CPU, this cpu is the active timer, since the other one will be gated */
-       if (cpu == 0) {
-               up_timer_enable();
-       }
+        if (cpu == 0 && g_cpugating_flag[cpu] == 1) {
+                up_timer_enable();
+        }
+
 }

```